+++ CHANGELOG.md
@@ -7,5 +7,6 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Lower many log levels in RabbitMQAdapter from INFO to DEBUG
 - Prevent pytest from collecting TestAdapter
 - Loosen dependencies